backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pyswisseph==2.10.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
httpx==0.27.0
python-dotenv==1.0.1
pytz==2024.1
pandas==2.2.2
astropy
kerykeion
flatlib
